add dotfile update confirmation

+++ scripts/utils.sh
@@ -9,17 +9,37 @@ gitpull () {
 }
 
 create_dotconfig_file () {
-  # create_config_file filename filename_in_message
+  # create local config files
   echo; echo "Creating $2 config file...";
   cp ~/.dotfiles/config/$1 ~/$1
   echo "~/$1 created!"
 }
 
 update_dotconfig_file () {
-
+  # overwrite local config with remote config
   if ! diff ~/$1 ~/.dotfiles/config/$1; then
-    cp ~/.dotfiles/config/$1 ~/$1
-    echo "~/$1 updated!\n"
+    echo "~/.dotfiles/config/$1 --> ~/$1? [y/n]"
+    read confirmation
+    if [ $confirmation = 'y' ]; then
+      cp ~/.dotfiles/config/$1 ~/$1
+      echo "~/$1 updated!\n"
+    else
+      echo "overwrite ~/$1 cancelled"
+    fi
+  fi
+}
+
+backupdate_dotconfig_file () {
+  # overwrite remote config with local config
+  if ! diff ~/.dotfiles/config/$1 ~/$1 ; then
+    echo "~/$1 --> ~/.dotfiles/config/$1? [y/n]"
+    read confirmation
+    if [ $confirmation = 'y' ]; then
+      cp ~/$1 ~/.dotfiles/config/$1
+      echo "~/.dotfiles/config/$1 updated!\n"
+    else
+      echo "overwrite ~/.dotfiles/config/$1 cancelled"
+    fi
   fi
 }
